rename: 폴더명 수정
폴더명에 -를 사용할 수 없는 문제로 모든 폴더를 약어로 변경
예시) admin -> adm
     personal-training -> ptr
