+++ src/main/resources/mapper/AdminMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org/DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 
 
-<mapper namespace="com.backend.hyunfit.domain.admin.mapper.AdminMapper">
+<mapper namespace="com.backend.hyunfit.domain.adm.mapper.AdminMapper">
 
     <select id="selectOneAdminByAdmId" parameterType="String" resultType="AdminDTO">
         SELECT adm_id, adm_pw
